requirements.txt
pymongo~=4.1.1
requests~=2.27.1
PyYAML~=6.0
pip~=22.0.4
platformdirs~=2.5.2
tomli~=2.0.1
pathspec~=0.9.0
click~=8.1.3
certifi~=2021.10.8
idna~=3.3
urllib3~=1.26.9
setuptools~=58.1.0